requirements.txt
Pillow==6.2.1
PyYAML==5.1.2
